backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
python-docx>=0.8.11
groq>=0.4.0
google-generativeai>=0.8.0
aiofiles>=23.0.0